docker-backend/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.4.1
boto3==1.21.9
botocore==1.24.9; python_version >= '3.6'
django-cors-headers==3.11.0
django-crispy-forms==1.13.0
django-filter==21.1
django-s3direct==2.0.0
django-storages==1.12.3
django==3.2.8
djangorestframework-simplejwt==5.0.0
djangorestframework==3.13.1
gunicorn==20.1.0
importlib-metadata==4.11.2; python_version < '3.10'
jmespath==0.10.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
markdown==3.3.6
pillow==8.4.0
psycopg2-binary==2.9.3
pyjwt==2.3.0; python_version >= '3.6'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2021.3
s3transfer==0.5.2; python_version >= '3.6'
setuptools==60.9.3; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2
urllib3==1.26.8; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
zipp==3.7.0; python_version >= '3.7'
